+++ commands/GemstonesOnFaceBetweenCurves/GemstonesOnFaceBetweenCurves.py
@@ -21,6 +21,7 @@ _sizeStepValueInput: adsk.core.ValueCommandInput = None
 _targetGapValueInput: adsk.core.ValueCommandInput = None
 _sizeRatioValueInput: adsk.core.ValueCommandInput = None
 _flipValueInput: adsk.core.BoolValueCommandInput = None
+_flipFaceNormalValueInput: adsk.core.BoolValueCommandInput = None
 _absoluteDepthOffsetValueInput: adsk.core.ValueCommandInput = None
 _relativeDepthOffsetValueInput: adsk.core.ValueCommandInput = None
 
@@ -91,8 +92,14 @@ sizeRatioInputDef = strings.InputDef(
 
 flipInputDef = strings.InputDef(
     'flip', 
-    'Flip Gemstone', 
+    'Flip Gemstones', 
     "Flip gemstone orientation.\nReverses the direction the gemstone faces relative to the surface."
+    )
+
+flipFaceNormalInputDef = strings.InputDef(
+    'flipFaceNormal',
+    'Flip Face Normal',
+    "Flip gemstone relative to face normal.\nRotates the gemstone 180 degrees around the face normal."
     )
 
 absoluteDepthOffsetInputDef = strings.InputDef(
@@ -183,7 +190,7 @@ class CreateCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
 
             global _faceSelectionInput, _curvesSelectionInput, _startOffsetValueInput, _endOffsetValueInput
             global _sizeStepValueInput, _targetGapValueInput, _sizeRatioValueInput
-            global _flipValueInput, _flipDirectionValueInput, _uniformDistributionValueInput, _absoluteDepthOffsetValueInput, _relativeDepthOffsetValueInput
+            global _flipValueInput, _flipFaceNormalValueInput, _flipDirectionValueInput, _uniformDistributionValueInput, _absoluteDepthOffsetValueInput, _relativeDepthOffsetValueInput
 
             _curvesSelectionInput = inputs.addSelectionInput(selectCurvesInputDef.id, selectCurvesInputDef.name, selectCurvesInputDef.tooltip)
             _curvesSelectionInput.addSelectionFilter(adsk.core.SelectionCommandInput.SketchCurves)
@@ -235,6 +242,10 @@ class CreateCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
             _flipValueInput = inputs.addBoolValueInput(flipInputDef.id, flipInputDef.name, True, '', flip)
             _flipValueInput.tooltip = flipInputDef.tooltip
 
+            flipFaceNormal = False
+            _flipFaceNormalValueInput = inputs.addBoolValueInput(flipFaceNormalInputDef.id, flipFaceNormalInputDef.name, True, '', flipFaceNormal)
+            _flipFaceNormalValueInput.tooltip = flipFaceNormalInputDef.tooltip
+
             absoluteDepthOffset = adsk.core.ValueInput.createByReal(0.0)
             _absoluteDepthOffsetValueInput = inputs.addValueInput(absoluteDepthOffsetInputDef.id, absoluteDepthOffsetInputDef.name, defaultLengthUnits, absoluteDepthOffset)
             _absoluteDepthOffsetValueInput.tooltip = absoluteDepthOffsetInputDef.tooltip
@@ -285,7 +296,7 @@ class EditCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
                 return
 
             global _startOffsetValueInput, _endOffsetValueInput, _sizeStepValueInput, _targetGapValueInput
-            global _sizeRatioValueInput, _flipValueInput, _flipDirectionValueInput, _uniformDistributionValueInput
+            global _sizeRatioValueInput, _flipValueInput,  _flipFaceNormalValueInput, _flipDirectionValueInput, _uniformDistributionValueInput
             global _absoluteDepthOffsetValueInput, _relativeDepthOffsetValueInput
 
             _curvesSelectionInput = inputs.addSelectionInput(selectCurvesInputDef.id, selectCurvesInputDef.name, selectCurvesInputDef.tooltip)
@@ -373,6 +384,14 @@ class EditCommandCreatedHandler(adsk.core.CommandCreatedEventHandler):
             _flipValueInput.tooltip = flipInputDef.tooltip
 
             try:
+                flipFaceNormalParam = params.itemById(flipFaceNormalInputDef.id)
+                flipFaceNormal = flipFaceNormalParam.expression.lower() == 'true'
+            except:
+                flipFaceNormal = False
+            _flipFaceNormalValueInput = inputs.addBoolValueInput(flipFaceNormalInputDef.id, flipFaceNormalInputDef.name, True, '', flipFaceNormal)
+            _flipFaceNormalValueInput.tooltip = flipFaceNormalInputDef.tooltip
+
+            try:
                 absoluteDepthOffsetParam = params.itemById(absoluteDepthOffsetInputDef.id)
                 absoluteDepthOffset = adsk.core.ValueInput.createByString(absoluteDepthOffsetParam.expression)
             except:
@@ -454,7 +473,7 @@ class ValidateInputsHandler(adsk.core.ValidateInputsEventHandler):
             if not all( [_startOffsetValueInput.isValidExpression, _endOffsetValueInput.isValidExpression,
                          _sizeStepValueInput.isValidExpression, _targetGapValueInput.isValidExpression,
                          _sizeRatioValueInput.isValidExpression,
-                         _flipValueInput.isValid,
+                         _flipValueInput.isValid, _flipFaceNormalValueInput.isValid,
                          _flipDirectionValueInput.isValid,
                          _uniformDistributionValueInput.isValid,
                          _absoluteDepthOffsetValueInput.isValidExpression,
@@ -513,6 +532,7 @@ class ExecutePreviewHandler(adsk.core.CommandEventHandler):
             targetGap = _targetGapValueInput.value
             sizeRatio = _sizeRatioValueInput.value
             flip = _flipValueInput.value
+            flipFaceNormal = _flipFaceNormalValueInput.value
             flipDirection = _flipDirectionValueInput.value
             uniformDistribution = _uniformDistributionValueInput.value
             absoluteDepthOffset = _absoluteDepthOffsetValueInput.value
@@ -530,10 +550,10 @@ class ExecutePreviewHandler(adsk.core.CommandEventHandler):
             baseFeature.startEdit()
 
             for point, size in pointsAndSizes:
-                gemstone = createGemstone(face, point, size, flip, absoluteDepthOffset, relativeDepthOffset)
+                gemstone = createGemstone(face, point, size, flip, absoluteDepthOffset, relativeDepthOffset, flipFaceNormal)
                 if gemstone is not None:
                     body = component.bRepBodies.add(gemstone, baseFeature)
-                    setGemstoneAttributes(body, flip, absoluteDepthOffset, relativeDepthOffset)
+                    setGemstoneAttributes(body, flip, absoluteDepthOffset, relativeDepthOffset, flipFaceNormal)
                     body.material = diamondMaterial
 
             baseFeature.finishEdit()
@@ -577,13 +597,13 @@ class CreateExecuteHandler(adsk.core.CommandEventHandler):
             baseFeature.startEdit()
 
             for point, size in pointsAndSizes:
-                gemstone = createGemstone(face, point, size, _flipValueInput.value, _absoluteDepthOffsetValueInput.value, _relativeDepthOffsetValueInput.value)
+                gemstone = createGemstone(face, point, size, _flipValueInput.value, _absoluteDepthOffsetValueInput.value, _relativeDepthOffsetValueInput.value, _flipFaceNormalValueInput.value)
                 if gemstone is None:
                     eventArgs.executeFailed = True
                     return
                 
                 body = comp.bRepBodies.add(gemstone, baseFeature)
-                setGemstoneAttributes(body, _flipValueInput.value, _absoluteDepthOffsetValueInput.value, _relativeDepthOffsetValueInput.value)
+                setGemstoneAttributes(body, _flipValueInput.value, _absoluteDepthOffsetValueInput.value, _relativeDepthOffsetValueInput.value, _flipFaceNormalValueInput.value)
                 body.material = diamondMaterial
 
             baseFeature.finishEdit()
@@ -621,6 +641,9 @@ class CreateExecuteHandler(adsk.core.CommandEventHandler):
                          
             flipInput = adsk.core.ValueInput.createByString(str(_flipValueInput.value).lower())
             customFeatureInput.addCustomParameter(flipInputDef.id, flipInputDef.name, flipInput, '', True)
+
+            flipFaceNormalInput = adsk.core.ValueInput.createByString(str(_flipFaceNormalValueInput.value).lower())
+            customFeatureInput.addCustomParameter(flipFaceNormalInputDef.id, flipFaceNormalInputDef.name, flipFaceNormalInput, '', True)
 
             absoluteDepthOffsetInput = adsk.core.ValueInput.createByString(_absoluteDepthOffsetValueInput.expression)
             customFeatureInput.addCustomParameter(absoluteDepthOffsetInputDef.id, absoluteDepthOffsetInputDef.name, absoluteDepthOffsetInput,
@@ -726,6 +749,11 @@ class EditExecuteHandler(adsk.core.CommandEventHandler):
             _editedCustomFeature.parameters.itemById(flipInputDef.id).expression = str(_flipValueInput.value).lower()
             _editedCustomFeature.parameters.itemById(absoluteDepthOffsetInputDef.id).expression = _absoluteDepthOffsetValueInput.expression
             _editedCustomFeature.parameters.itemById(relativeDepthOffsetInputDef.id).expression = _relativeDepthOffsetValueInput.expression
+            
+            try:
+                _editedCustomFeature.parameters.itemById(flipFaceNormalInputDef.id).expression = str(_flipFaceNormalValueInput.value).lower()
+            except:
+                pass
 
         except:
             showMessage(f'EditExecuteHandler: {traceback.format_exc()}\n', True)
@@ -806,6 +834,11 @@ def updateFeature(customFeature: adsk.fusion.CustomFeature) -> bool:
         
         flip = customFeature.parameters.itemById(flipInputDef.id).expression.lower() == 'true'
         
+        try:
+            flipFaceNormal = customFeature.parameters.itemById(flipFaceNormalInputDef.id).expression.lower() == 'true'
+        except:
+            flipFaceNormal = False
+        
         absoluteDepthOffset = customFeature.parameters.itemById(absoluteDepthOffsetInputDef.id).value
         relativeDepthOffset = customFeature.parameters.itemById(relativeDepthOffsetInputDef.id).value
 
@@ -820,18 +853,18 @@ def updateFeature(customFeature: adsk.fusion.CustomFeature) -> bool:
 
             if i < baseFeature.bodies.count:
                 currentBody = baseFeature.bodies.item(i)
-                newBody = updateGemstone(currentBody, faceEntity, point, size, flip, absoluteDepthOffset, relativeDepthOffset)
+                newBody = updateGemstone(currentBody, faceEntity, point, size, flip, absoluteDepthOffset, relativeDepthOffset, flipFaceNormal)
                 if newBody is not None:
                     baseFeature.updateBody(currentBody, newBody)
                 else:
                     success = False
             else:
-                gemstone = createGemstone(faceEntity, point, size, flip, absoluteDepthOffset, relativeDepthOffset)
+                gemstone = createGemstone(faceEntity, point, size, flip, absoluteDepthOffset, relativeDepthOffset, flipFaceNormal)
                 if gemstone is not None:
                     body = component.bRepBodies.add(gemstone, baseFeature)
                     body.material = diamondMaterial
                     
-                    if not _isRolledForEdit: setGemstoneAttributes(body, flip, absoluteDepthOffset, relativeDepthOffset)
+                    if not _isRolledForEdit: setGemstoneAttributes(body, flip, absoluteDepthOffset, relativeDepthOffset, flipFaceNormal)
                     
                 else:
                     success = False